news/requirements.txt
feedparser
requests
beautifulsoup4
python-dotenv
pydantic
openai